Added WindowsJNLPData
This allows a way of attaching Windows JNLP inbound agents that are dynamically provisioned. This plugin spins up the ec2 instances as needed. The agent name and key needed to connect to the controller are added as tags on the instance. The instance is responsible for in the userdata script retrieving the pertinent values from tags and then starting up an agent process.

This could in theory be also used for non-windows agents. However, the ssh protocol agents seem to work quite nicely. We won't be on windows builders much longer so this should suffice.